+++ common.py
@@ -44,12 +44,6 @@ AV_DELETE_INFECTED_FILES = os.getenv("AV_DELETE_INFECTED_FILES", "False")
 AV_DEFINITION_FILE_PREFIXES = ["main", "daily", "bytecode"]
 AV_DEFINITION_FILE_SUFFIXES = ["cld", "cvd"]
 
-IS_AV_ENABLED = os.getenv("IS_AV_ENABLED", "True")
-MIME_VALIDATION_NONE = "no-validation"
-MIME_VALIDATION_STATIC = "static"
-MIME_VALIDATION_S3_CONTENT_TYPE = "s3-content-type"
-MIME_VALIDATION = os.getenv("MIME_VALIDATION", MIME_VALIDATION_STATIC)
-MIME_VALIDATION_STATIC_VALID_LIST = os.getenv("VALID_MIMES", "image/gif,image/png,image/jpeg,image/jpg,application/pdf")
 
 def create_dir(path):
     if not os.path.exists(path):
